segregate functions calling order_command() into /bin/src/ctl/request_builder.rs
